# docs/swagger.yaml
basePath: /
definitions:
  request.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
  request.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  response.AuthResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/response.UserResponse'
    type: object
  response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.UserPaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.UserResponse'
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
  response.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  utils.Response:
    properties:
      data:
        description: 'omitempty: omitted if empty'
      errors:
        description: Used for specific error details/validation
      message:
        type: string
      meta:
        description: Used for pagination metadata
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an inventory management system API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Inventory System API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, and return a stateful
        UUID session token.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.AuthResponse'
              type: object
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: User Login
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user by revoking their current UUID session.
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: User Logout
      tags:
      - Auth
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: |-
        Retrieve a paginated list of users with optional search filtering.
        **Required Roles:** `super_admin`, `admin`
      parameters:
      - description: 'Page number for pagination (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Search filter for user name or email
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.UserPaginatedResponse'
              type: object
        "401":
          description: Unauthorized - Invalid or expired session
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: |-
        Register a new user (Admin or Staff) into the system. Requires authentication.
        **Required Roles:** `super_admin`, `admin`
      parameters:
      - description: User data payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your UUID token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"