# internal/handler/auth_hdl.go
package handler

import (
	"encoding/json"
	"net/http"
	"strings"

	"inventory-system/internal/dto/request"
	"inventory-system/internal/service"
	"inventory-system/pkg/utils"

	"github.com/go-chi/chi/v5/middleware"
	"go.uber.org/zap"
)

// AuthHandler is the concrete implementation of AuthHandler.
type AuthHandler struct {
	authService service.AuthService
	logger      *zap.Logger
}

// NewAuthHandler creates and returns a new instance of AuthHandler.
func NewAuthHandler(authService service.AuthService, logger *zap.Logger) *AuthHandler {
	return &AuthHandler{
		authService: authService,
		logger:      logger,
	}
}

// Login godoc
// @Summary      User Login
// @Description  Authenticate user with email and password, and return a stateful UUID session token.
// @Tags         Auth
// @Accept       json
// @Produce      json
// @Param        request body request.LoginRequest true "Login credentials"
// @Success      200  {object}  utils.Response{data=response.AuthResponse} "Login successful"
// @Failure      400  {object}  utils.Response "Invalid request format"
// @Failure      401  {object}  utils.Response "Invalid email or password"
// @Failure      500  {object}  utils.Response "Internal server error"
// @Router       /api/v1/auth/login [post]
func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
	// Extract the unique Request ID generated by the Chi middleware.
	// This ID is used for tracing the request across the application layers.
	reqID := middleware.GetReqID(r.Context())

	var req request.LoginRequest

	// 1. Decode the JSON payload from the request body into the DTO.
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		h.logger.Warn("Failed to decode login request body",
			zap.String("request_id", reqID),
			zap.Error(err),
		)
		utils.Error(w, r, http.StatusBadRequest, "Invalid request format", nil)
		return
	}

	// 2. Pass the decoded request to the Service layer for business logic processing.
	res, err := h.authService.Login(r.Context(), req)
	if err != nil {
		// Handle expected business errors (e.g., wrong credentials)
		if err.Error() == "invalid email or password" {
			h.logger.Warn("Unauthorized login attempt", zap.String("request_id", reqID), zap.String("email", req.Email))
			utils.Error(w, r, http.StatusUnauthorized, err.Error(), nil)
			return
		}

		// Handle unexpected system errors (e.g., database down, JWT generation failed)
		h.logger.Error("Internal server error during login", zap.String("request_id", reqID), zap.Error(err))
		utils.Error(w, r, http.StatusInternalServerError, "Internal server error", nil)
		return
	}

	// 3. If successful, return the HTTP 200 OK response with the generated token and user data.
	h.logger.Info("Login request processed successfully", zap.String("request_id", reqID))
	utils.Success(w, r, http.StatusOK, "Login successful", res)
}

// Logout godoc
// @Summary      User Logout
// @Description  Logout user by revoking their current UUID session.
// @Tags         Auth
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object}  utils.Response "Logout successful"
// @Failure      401  {object}  utils.Response "Unauthorized"
// @Failure      500  {object}  utils.Response "Internal server error"
// @Router       /api/v1/auth/logout [post]
func (h *AuthHandler) Logout(w http.ResponseWriter, r *http.Request) {
	// 1. Ekstrak token dari header "Authorization: Bearer <token>"
	authHeader := r.Header.Get("Authorization")
	if authHeader == "" {
		utils.Error(w, r, http.StatusUnauthorized, "Missing authorization header", nil)
		return
	}

	// Pisahkan "Bearer" dan token aslinya
	parts := strings.Split(authHeader, " ")
	if len(parts) != 2 || parts[0] != "Bearer" {
		utils.Error(w, r, http.StatusUnauthorized, "Invalid authorization format", nil)
		return
	}

	tokenString := parts[1]

	// 2. Lempar ke Service
	err := h.authService.Logout(r.Context(), tokenString)
	if err != nil {
		utils.Error(w, r, http.StatusInternalServerError, err.Error(), nil)
		return
	}

	// 3. Kembalikan response sukses
	utils.Success(w, r, http.StatusOK, "Logout successful", nil)
}